refactor-HomeScreen 구조 변경
