+++ src/screens/HomeScreen/HomeScreenStyles.ts
@@ -1,0 +1,73 @@
+import {StyleSheet, Dimensions} from 'react-native';
+const {width, height} = Dimensions.get('window');
+import {FONT} from '@/src/constants/styles/font/default-font';
+
+export const HomeScreenStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F5F5F5',
+    width: width,
+    height: height,
+  },
+
+  profileTextContainer: {
+    marginLeft: 130,
+    marginTop: -30,
+    flexDirection: 'row',
+  },
+
+  modalHeaderContainer: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    width: '100%',
+    marginBottom: height * 0.02,
+  },
+  modalOverlay: {
+    flex: 1,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  overlayTouchable: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+  },
+  modalView: {
+    width: width * 0.8,
+    maxHeight: height * 0.6,
+    backgroundColor: 'white',
+    borderRadius: 8,
+    padding: width * 0.05,
+    alignItems: 'center',
+    shadowColor: '#000',
+    shadowOffset: {
+      width: 0,
+      height: 2,
+    },
+    shadowOpacity: 0.25,
+    shadowRadius: 4,
+    elevation: 5,
+  },
+  modalText: {
+    marginBottom: height * 0.02,
+    textAlign: 'left',
+    fontSize: width * 0.04,
+    fontWeight: '700',
+    fontFamily: `${FONT}`,
+    color: '#000000',
+  },
+  closeButton: {
+    backgroundColor: '#1AA5AA',
+    borderRadius: 4,
+    paddingVertical: height * 0.015,
+    paddingHorizontal: width * 0.1,
+  },
+  closeButtonText: {
+    color: 'white',
+    fontWeight: 'bold',
+    fontSize: width * 0.035,
+  },
+});
